refatorando Locker

+++ src/main/java/br/com/inatel/sherlock/repository/LockerRepository.java
@@ -3,5 +3,5 @@ package br.com.inatel.sherlock.repository;
 import br.com.inatel.sherlock.models.Locker;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LockerRepository extends JpaRepository<Locker, Integer> {
+public interface LockerRepository extends JpaRepository<Locker, Long> {
 }
